+++ changelog.md
@@ -1,2 +1,7 @@
 # Version 13.3 on 13-07-26
 Finally dropping the SNAPSHOT before extracting experimentation code into libexp.
+
+# Version 14
+- drop some dependencies (libtreewidth, logging)
+- migrate to scalaz 7
+- migrate to specs2 2.1